[K32W0] Move FactoryDataProvider to K32W common area
Update #include statements to use an absolute path.
